% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions}
\alias{actions}
\alias{docker_cli}
\alias{rscript_byod}
\alias{filter}
\alias{rsync}
\alias{ghpages}
\alias{netlify}
\title{R wrappers around GitHub actions}
\usage{
docker_cli(IDENTIFIER = "Build Image",
  args = "build --tag=repo:latest .")

rscript_byod(IDENTIFIER = "Arbitrary Rscript", needs, args)

filter(IDENTIFIER = "Filter", needs, args = "branch master")

rsync(IDENTIFIER = "Rsync", needs, env, args)

ghpages(IDENTIFIER = "Deploy to GitHub Pages", needs,
  env = "BUILD_DIR = 'public/'")

netlify(IDENTIFIER = "Deploy to Netlify", needs,
  args = "deploy --dir=site --functions=functions")
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action or workflow block.
Used:
\itemize{
\item as informative label on GitHub.com,
\item in the \code{needs} fields of other \emph{action blocks} to express dependencies.
\item in the \code{resolves} fields of other \emph{workflow blocks} to express dependencies.
}}

\item{args}{\code{[character()]}
giving the arguments to pass to the action.
Arguments get appended to the last command in \code{ENTRYPOINT}.
Defaults to \code{NULL} for no arguments.}

\item{needs}{\code{[character()]}
giving the actions (by their \code{IDENTIFIER}s) that must complete successfully before this action will be invoked.
Defaults to \code{NULL} for no upstream dependencies.}

\item{env}{\code{[list(character(1)]}
giving the environment variables to set in the action's runtime environment.
Defaults to \code{NULL} for no environment variables.}
}
\description{
These functions are for \strong{advanced users} knowledgeable about GitHub actions.
Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories.
For documentation on these actions, consult their respective \code{README.md}s linked in the below.
Some of these action wrappers include sensible defaults for most uses in R.
You can always create action blocks entirely from scratch using \code{\link[=make_action_block]{make_action_block()}}.

The \code{uses} field is always hardcoded to a particular commit or tag of the underlying github action to ensure.
}
\section{Functions}{
\itemize{
\item \code{docker_cli}: \href{https://github.com/actions/docker/tree/aea64bb1b97c42fa69b90523667fef56b90d7cff}{Docker CLI}

\item \code{rscript_byod}: \href{https://github.com/maxheld83/ghactions/tree/master/Rscript-byod}{Rscript-byod}

\item \code{filter}: \href{https://github.com/actions/bin/tree/a9036ccda9df39c6ca7e1057bc4ef93709adca5f/filter}{filter}

\item \code{rsync}: \href{https://github.com/maxheld83/rsync/tree/v0.1.1}{rsync}

\item \code{ghpages}: \href{https://github.com/maxheld83/ghpages/tree/v0.1.2}{ghpages}

\item \code{netlify}: \href{https://github.com/netlify/actions/tree/645ae7398cf5b912a3fa1eb0b88618301aaa85d0/cli/}{netlify}
}}

\concept{actions}
