% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{browse_github_actions}
\alias{browse_github_actions}
\title{Quickly browse to important package webpages}
\usage{
browse_github_actions(package = NULL)
}
\arguments{
\item{package}{Name of package; leave as \code{NULL} to use current package}
}
\description{
Visits the GitHub actions page.
See \code{\link[usethis:browse_github]{usethis::browse_github()}} for details.
}
\seealso{
Other setup: \code{\link{edit_workflow}},
  \code{\link{use_dockerfile}},
  \code{\link{use_ghactions_badge}},
  \code{\link{use_ghactions}}
}
\concept{setup}
