% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{checkout}
\alias{checkout}
\title{Create a step to checkout a repository}
\usage{
checkout()
}
\description{
Create a step to checkout a repository
}
\seealso{
Other steps: \code{\link{firebase}}, \code{\link{ghpages}},
  \code{\link{install_deps}}, \code{\link{netlify}},
  \code{\link{pkg_dev}}, \code{\link{rscript}},
  \code{\link{rsync}}

Other installation: \code{\link{install_deps}}
}
\concept{installation}
\concept{steps}
