% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{container}
\alias{container}
\title{Create nested list for the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idcontainer}{container} field in \code{\link[=job]{job()}}}
\usage{
container(image, env = NULL, ports = NULL, volumes = NULL,
  options = NULL)
}
\arguments{
\item{image}{\code{[character(1)]}
giving the published docker image to use as the container to run the action.}

\item{env}{\code{[list()]}
giving environment variables for the container as a \emph{named} list.
Defaults to \code{NULL}.}

\item{ports, volumes}{\code{[list()]}
giving ports to expose, and volumes for the container to use as an \emph{unnamed} list.
Defaults to \code{NULL}.}

\item{options}{\code{[character()]}
giving additional options.
Defaults to \code{NULL}.}
}
\description{
Create nested list for the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idcontainer}{container} field in \code{\link[=job]{job()}}
}
\seealso{
Other syntax: \code{\link{gh_matrix}},
  \code{\link{ghactions_events}},
  \code{\link{ghactions_vms}}, \code{\link{job}},
  \code{\link{on}}, \code{\link{strategy}},
  \code{\link{workflow}}
}
\concept{syntax}
