% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{docker_cli}
\alias{docker_cli}
\title{Docker CLI}
\usage{

  docker_cli(uses = "actions/docker/cli@c08a5fc9e0286844156fefff2c141072048141f6",
  args = "build --tag=repo:latest .")
}
\arguments{
\item{uses}{\code{[character(1)]}
giving the Docker image that will run the action.}

\item{args}{\code{[character()]}
giving the arguments to pass to the action.
Arguments get appended to the last command in \code{ENTRYPOINT}.
Defaults to \code{NULL} for no arguments.}
}
\description{
Minimal wrapper around the (external) \href{https://github.com/actions/docker/tree/master/cli}{Docker CLI} action.
Below details are reproduced from the original Docker repository.
}
\seealso{
Other actions: \code{\link{filter}},
  \code{\link{rscript_byod}}, \code{\link{rsync}}
}
\concept{actions}
