% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{edit_workflow}
\alias{edit_workflow}
\title{Open configuration files}
\usage{
edit_workflow()
}
\description{
Open \code{.github/workflows/main} configuration file for GitHub actions.
See \code{\link[usethis:edit]{usethis::edit()}} for details.
}
\seealso{
Other setup: \code{\link{is_github_actions}},
  \code{\link{use_ghactions_badge}},
  \code{\link{use_ghactions}}
}
\concept{setup}
