% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{filter}
\alias{filter}
\title{Filter}
\usage{

  filter(uses = "actions/bin/filter@c6471707d308175c57dfe91963406ef205837dbd",
  needs = "Render", args = "branch master")
}
\arguments{
\item{uses}{\code{[character(1)]}
giving the Docker image that will run the action.}

\item{needs}{\code{[character()]}
giving the actions (by their \code{IDENTIFIER}s) that must complete successfully before this action will be invoked.
Defaults to \code{NULL} for no upstream dependencies.}

\item{args}{\code{[character()]}
giving the arguments to pass to the action.
Arguments get appended to the last command in \code{ENTRYPOINT}.
Defaults to \code{NULL} for no arguments.}
}
\description{
Filter
}
\seealso{
Other actions: \code{\link{docker_cli}},
  \code{\link{rscript_byod}}, \code{\link{rsync}}
}
\concept{actions}
