% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{gh_matrix}
\alias{gh_matrix}
\title{Create nested list for the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idstrategy}{matrix} field in \code{\link[=strategy]{strategy()}}}
\usage{
gh_matrix(..., exclude = NULL, include = NULL)
}
\arguments{
\item{...}{\code{[character()]}
giving values for variable for the matrix build.}

\item{exclude, include}{\code{[list(list(character(1)))]}
giving unnamed lists of combinations of variables to ex- or include.
Defaults to \code{NULL}.}
}
\description{
Create nested list for the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idstrategy}{matrix} field in \code{\link[=strategy]{strategy()}}
}
\seealso{
Other syntax: \code{\link{container}},
  \code{\link{ghactions_events}},
  \code{\link{ghactions_vms}}, \code{\link{job}},
  \code{\link{on}}, \code{\link{step}},
  \code{\link{strategy}}, \code{\link{workflow}}
}
\concept{syntax}
