% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghactions-package.R
\docType{package}
\name{ghactions}
\alias{ghactions}
\alias{ghactions-package}
\title{ghactions: GitHub actions for R}
\description{
Provides templating functions with sensible defaults to use GitHub Actions for R projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/ghactions}
  \item Report bugs at \url{https://github.com/r-lib/ghactions/issues}
}

}
\author{
\strong{Maintainer}: Max Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Authors:
\itemize{
  \item Jim Hester \email{james.f.hester@gmail.com} (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})
}

Other contributors:
\itemize{
  \item Verena Held \email{info@verenaheld.in} [contributor]
  \item RStudio Inc. [copyright holder, funder]
}

}
\keyword{internal}
