% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\docType{data}
\name{ghactions_events}
\alias{ghactions_events}
\title{Supported events to trigger GitHub actions}
\format{
An object of class \code{character} of length 28.
}
\usage{
ghactions_events
}
\description{
You can trigger GitHub actions from these events.
List is taken from \href{https://developer.github.com/actions/creating-workflows/workflow-configuration-options/#events-supported-in-workflow-files}{official spec}.
}
\examples{
ghactions_events

}
\seealso{
Other syntax: 
\code{\link{container}()},
\code{\link{gh_matrix}()},
\code{\link{ghactions_vms}},
\code{\link{job}()},
\code{\link{on}()},
\code{\link{step}()},
\code{\link{strategy}()},
\code{\link{workflow}()}
}
\concept{syntax}
\keyword{datasets}
