% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\docType{data}
\name{ghactions_vms}
\alias{ghactions_vms}
\title{Virtual machines \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idruns-on}{available} on GitHub Actions}
\format{An object of class \code{character} of length 8.}
\usage{
ghactions_vms
}
\description{
Virtual machines \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idruns-on}{available} on GitHub Actions
}
\examples{
ghactions_vms

}
\seealso{
Other syntax: \code{\link{ghactions_events}},
  \code{\link{io}}, \code{\link{job}}, \code{\link{on}},
  \code{\link{workflow}}
}
\concept{syntax}
\keyword{datasets}
