% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{install_deps}
\alias{install_deps}
\title{Create a step to install R package dependencies}
\usage{
install_deps(name = "Install Package Dependencies", ...)
}
\arguments{
\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
}}
}
\description{
Installs R package dependencies from a \code{DESCRIPTION} at the repository root.
}
\concept{steps actions installation}
