% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_github_actions}
\alias{is_github_actions}
\title{Test whether runtime is GitHub actions}
\usage{
is_github_actions()
}
\description{
Test whether runtime is GitHub actions
}
\seealso{
Other setup: \code{\link{edit_workflow}},
  \code{\link{use_ghactions_badge}},
  \code{\link{use_ghactions}}
}
\concept{setup}
