% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{make_action}
\alias{make_action}
\title{Create GitHub Actions syntax for one action.}
\usage{
make_action(IDENTIFIER, needs = NULL, uses, runs = NULL, args = NULL,
  env = NULL, secrets = NULL)
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action.
Shown on github.com and used in the \code{needs} fields of other actions.}

\item{needs}{\code{[character()]}
giving the actions (by their \code{IDENTIFIER}s) that must complete successfully before this action will be invoked.
Defaults to \code{NULL} for no upstream dependencies.}

\item{uses}{\code{[character(1)]}
giving the Docker image that will run the action.}

\item{runs}{\code{[character(1)]}
giving the command to run in the docker image.
Overrides the \code{Dockerfile} \code{ENTRYPOINT}.
Defaults to \code{NULL} for no commands (recommended).}

\item{args}{\code{[character()]}
giving the arguments to pass to the action.
Arguments get appended to the last command in \code{ENTRYPOINT}.
Defaults to \code{NULL} for no arguments.}

\item{env}{\code{[list(character(1)]}
giving the environment variables to set in the action's runtime environment.
Defaults to \code{NULL} for no environment variables.}

\item{secrets}{\code{[character()]}
giving the \emph{names} of the secret variables to set in the runtime enviornment, which the action can access as an environment variable.
The \emph{values} of secrets must be set in your repository's "Settings" tab.
\strong{Do not store secrets in your repository.}
GitHub advises against using GitHub actions for production secrets during the public beta period.
Defaults to \code{NULL} for no secrets.}
}
\value{
\code{[character(1)]}
}
\description{
Creates the syntax for \emph{one} GitHub action as a string.
For details on the arguments, see \href{https://developer.github.com/actions/creating-workflows/workflow-configuration-options/}{here}.-
}
\details{
The \code{main.workflow} files used in \href{http://github.com/features/actions}{GitHub Actions} is comprised of several actions.
}
\examples{
make_action(
  IDENTIFIER = "Simple Addition",
  uses = "maxheld83/ghactions/Rscript-byod@master",
  needs = "Build Image",
  args = "-e '1+1'"
)

}
\keyword{internal}
