% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{make_template}
\alias{make_template}
\title{Fill in template}
\usage{
make_template(l, template)
}
\arguments{
\item{l}{\code{[list()]}
giving the named list of HCL fields.}

\item{template}{\code{[character(1)]}
giving the name of the template file.}
}
\value{
\code{[character()]}
of class \link[glue:glue]{glue::glue}, giving the syntax for one workflow or action block.
}
\description{
Fill in template
}
\keyword{internal}
