% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{netlify}
\alias{netlify}
\title{Create an action step to deploy to \href{https://www.netlify.com}{Netlify}}
\usage{
netlify(name = "Deploy to Netlify",
  `if` = "github.ref == 'refs/heads/master'", dir = "_site",
  prod = TRUE, with = NULL, env = NULL, site, ...)
}
\arguments{
\item{dir}{\code{[character(1)]}
giving the path relative from your \code{/github/workspace} to the directory to be published.}

\item{prod}{\code{[logical(1)]}
giving whether the deploy should be to production.}

\item{site}{\code{[character(1)]}
giving a site ID to deploy to.}

\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
}}
}
\description{
\strong{Remember to provide \code{NETLIFY_AUTH_TOKEN} and \code{NETLIFY_SITE_ID} (optional) as secrets to the GitHub UI.}
}
\details{
Wraps the external \href{https://github.com/netlify/actions}{netlify cli action}.
}
\seealso{
Other steps actions deployment: \code{\link{firebase}},
  \code{\link{ghpages}}, \code{\link{rsync}}
}
\concept{steps actions deployment}
