% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{on}
\alias{on}
\title{Create nested list for an \code{on:} field.}
\usage{
on(event, ...)
}
\arguments{
\item{event}{\code{[character(1)]}
giving the event on which to filter.
Must be \emph{one} of \code{c("push", "pull_request", "schedule")}.}

\item{...}{\code{[character()]}
giving the filters on which to run
Must correspond to the filters allowed by \code{event}.}
}
\description{
Create nested list for an \code{on:} field.
}
\details{
See the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions}{GitHub Actions workflow syntax} for details.
}
\examples{
on(
  event = "push",
  branches = c("master", "releases/*")
)
}
\seealso{
Other syntax: \code{\link{ghactions_events}},
  \code{\link{ghactions_vms}}, \code{\link{io}},
  \code{\link{job}}, \code{\link{workflow}}
}
\concept{syntax}
