% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{pkg_dev}
\alias{pkg_dev}
\alias{rcmd_check}
\alias{covr}
\title{CI/CD steps for a package at the repository root}
\usage{
rcmd_check(name = "Check Package")

covr(name = "Run Code Coverage")
}
\arguments{
\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
}}
}
\description{
CI/CD steps for a package at the repository root
}
\section{Functions}{
\itemize{
\item \code{rcmd_check}: \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}

\item \code{covr}: \code{\link[covr:codecov]{covr::codecov()}}
}}

\concept{steps pkg_development}
