% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{rcmd_check}
\alias{rcmd_check}
\title{Create a step to check an R package}
\usage{
rcmd_check(name = "Check Package")
}
\arguments{
\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
}}
}
\description{
Runs \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} on a package at the repository root.
}
\concept{steps pkg_development}
