% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{read_workflows}
\alias{read_workflows}
\alias{read_workflow}
\title{Read in GitHub Actions workflows}
\usage{
read_workflows(path = ".github/workflows")

read_workflow(path)
}
\arguments{
\item{path}{\code{[character()]} giving the directory from the repository root where to find GitHub Actions workflows.
Defaults to \code{".github/workflows"}.}
}
\value{
\code{[list()]} of lists from yaml.
}
\description{
Read in GitHub Actions workflows
}
\section{Functions}{
\itemize{
\item \code{read_workflow}: Read in an individual GitHub Actions workflow
}}

\concept{workflows}
