% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{readme2sections}
\alias{readme2sections}
\title{Create roxygen sections from GitHub Action Readmes}
\usage{
readme2sections(path = "Rscript-byod/README.md")
}
\arguments{
\item{path}{\code{[character(1)]}
Giving path from repo root.}
}
\description{
Create roxygen sections from GitHub Action Readmes
}
