% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{edit_workflow}
\alias{browse_github_actions}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{usethis}{\code{\link[usethis]{edit_workflow}}, \code{\link[usethis]{browse_github_actions}}}
}}

