% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{rscript}
\alias{rscript}
\title{Create a step to run \link[utils:Rscript]{utils::Rscript}}
\usage{
rscript(options = NULL, expr = NULL, file = NULL, args = NULL, ...)
}
\arguments{
\item{options}{a list of options, all beginning with \samp{--}.  These
    can be any of the options of the standard \R front-end, and also those
    described in the details.}

\item{expr}{\R expression(s), properly quoted.}

\item{file}{the name of a file containing \R commands.  \samp{-}
    indicates \file{stdin}.}

\item{args}{arguments to be passed to the script in \code{file}.}

\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{uses}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{working-directory}{\code{[character(1)]}
giving the default working directory.
Defaults to \code{NULL}.}
  \item{continue-on-error}{\code{[logical(1)]}
giving whether to allow a job to pass when this step fails.
Defaults to \code{NULL}.}
  \item{timeout-minutes}{\code{[integer(1)]}
giving the maximum number of minutes to run the step before killing the process.
Defaults to \code{NULL}.}
}}
}
\description{
Create a step to run \link[utils:Rscript]{utils::Rscript}
}
\seealso{
Other steps: \code{\link{checkout}},
  \code{\link{firebase}}, \code{\link{ghpages}},
  \code{\link{install_deps}}, \code{\link{netlify}},
  \code{\link{pkg_dev}}, \code{\link{rsync}}
}
\concept{script}
\concept{steps}
