% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{rscript_byod}
\alias{rscript_byod}
\title{Rscript \emph{bring-your-own-dockerfile}}
\usage{
rscript_byod(uses = "maxheld83/ghactions/Rscript-byod@master",
  needs = "Build image", args)
}
\arguments{
\item{uses}{\code{[character(1)]}
giving the Docker image that will run the action.}

\item{needs}{\code{[character()]}
giving the actions (by their \code{IDENTIFIER}s) that must complete successfully before this action will be invoked.
Defaults to \code{NULL} for no upstream dependencies.}

\item{args}{\code{[character()]}
giving the arguments to pass to the action.
Arguments get appended to the last command in \code{ENTRYPOINT}.
Defaults to \code{NULL} for no arguments.}
}
\description{
Minimal wrapper around the (internal) \href{https://github.com/maxheld83/ghactions/tree/master/Rscript-byod}{Rscript-byod} action.
}
\details{
This is a very \strong{simple} GitHub action; it just runs in \code{Rscript} whatever you provide as the \code{args} field (see below).
This action is meant for \strong{generic} R projects with \strong{arbitrary build environments}; you therefore \strong{have to bring your own dockerfile}.
If you are targeting a specific build or runtime environment (say, \code{R CMD build} or \href{https://www.shinyapps.io}{shinyapps.io}), some of the other actions may be more suitable for your purposes.
If you are \emph{not} targeting a specific environment, and don't know much about docker, you can easily start with one of the popular rocker images as in the below example.
Learn more about docker and R from the \href{http://rocker-project.org}{rocker project}.

A dockerfile is simply a text file called \code{Dockerfile} at the root of your repository.
At a minimum, it should include a \code{FROM} statement as in the below.
You are highly recommended to always use versioned images.\preformatted{FROM rocker/verse:3.5.2
}

Because a dockerfile is just a \emph{recipe} for a build environment, \strong{you first have to build your \code{repo/Dockerfile}} into an image.
Happily, GitHub actions already provides an \href{https://github.com/actions/docker}{action for that}, which you should run as a first step in your \code{main.workflow} with minimal customization:\preformatted{action "Build image" build --tag=repo:latest .
}

This just runs \code{docker build --tag=repo:latest .}, where \code{.} is the root of your repository.
Docker will then pick whichever \code{Dockerfile} it finds there, and well, \emph{build} it.
Your dockerfile recipe has now been prepared into a meal, and this meal (the docker \emph{image}) now exists in your \code{/github/workspace}.
This is a special directory that you won't ever see anywhere.
GitHub actions provisions this directory, and lets it \emph{persist as long as your \code{main.workflow} runs}.

Any downstream actions (such as \emph{this} action!) can now use the prepared image, but they have to know its name.
The \code{--tag=repo:latest} part of the above call simply names your image, literally "repo:latest".
This is just my convention, not a magic name.
But it is very important you \emph{use exactly this name}, because otherwise downstream actions (\emph{this} action) cannot base their own work on it.
This isn't terribly elegant, but currently appears to be the only way on GitHub actions to identify images from past actions (see \href{https://github.com/maxheld83/ghactions/issues/1}{this issue}).

Once your "bring-your-own-dockerfile" has been build into an image, the present action then simply runs whatever you pass \code{Rscript} \emph{in} that image.Environment Variables

None so far.Secrets

nada.Required Arguments

Whatever your provide in \code{args} simply gets appended to the \code{Rscript} call.
So just image \code{Rscript} in front of it; if it is a valid shell command, you're good to go.
Remember that \code{Rscript} is the (\href{https://stackoverflow.com/questions/18306362/run-r-script-from-command-line/18306656#18306656}{now preferred}) way to run R from a system shell (\emph{not} the R console).
We have to use \code{Rscript}, because there's no GUI that could hook us up directly to the R console in our non-interactive image run.
\href{https://stat.ethz.ch/R-manual/R-devel/library/utils/html/Rscript.html}{Here} is the full documentation.

A typical \code{args} example is\preformatted{"-e 'rmarkdown::render_site()'"
}

which is appended to \code{Rscript} by this action, so you're really running\preformatted{Rscript -e 'rmarkdown::render_site()'
}

which \code{Rscript} passes to the R console, where you know it as\preformatted{rmarkdown::render_site()
}Optional Arguments

Nope.Example Usage

Here is a complete \code{main.workflow} that first builds your (byod) image, and then runs \code{rmarkdown::render_site()} on it.\preformatted{workflow "Build then render" Render rmarkdown

action "Build image" build --tag=repo:latest .

action "Render rmarkdown" -e 'rmarkdown::render_site()'
}

Of course, you can also use the visual workflow editor in GitHub, which is highly recommended.

The cryptic-seeming \code{uses = "maxheld83/ghactions/Rscript-byod@master"} in the above is simply a way to take a \emph{particular} version of this action as a dependency.
The part after \code{@} is a git tag, or release from my (= the action developer's) repo.
It is generally recommended to specify dependencies in GitHub actions as narrowly as you can.
That way, your workflow won't break if the maintainer of the action changes things up.
}
\seealso{
Other actions: \code{\link{docker_cli}},
  \code{\link{filter}}, \code{\link{rsync}}
}
\concept{actions}
