% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{rsync}
\alias{rsync}
\title{Rsync}
\usage{
rsync(uses = "maxheld83/rsync@v0.1.1", needs = "Master",
  secrets = c("SSH_PRIVATE_KEY", "SSH_PUBLIC_KEY"), env, args)
}
\arguments{
\item{uses}{\code{[character(1)]}
giving the Docker image that will run the action.}

\item{needs}{\code{[character()]}
giving the actions (by their \code{IDENTIFIER}s) that must complete successfully before this action will be invoked.
Defaults to \code{NULL} for no upstream dependencies.}

\item{secrets}{\code{[character()]}
giving the \emph{names} of the secret variables to set in the runtime enviornment, which the action can access as an environment variable.
The \emph{values} of secrets must be set in your repository's "Settings" tab.
\strong{Do not store secrets in your repository.}
GitHub advises against using GitHub actions for production secrets during the public beta period.
Defaults to \code{NULL} for no secrets.}

\item{env}{\code{[list(character(1)]}
giving the environment variables to set in the action's runtime environment.
Defaults to \code{NULL} for no environment variables.}

\item{args}{\code{[character()]}
giving the arguments to pass to the action.
Arguments get appended to the last command in \code{ENTRYPOINT}.
Defaults to \code{NULL} for no arguments.}
}
\description{
Rsync
}
\seealso{
Other actions: \code{\link{docker_cli}},
  \code{\link{filter}}, \code{\link{rscript_byod}}
}
\concept{actions}
