% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{strategy}
\alias{strategy}
\title{Create nested list for the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idstrategy}{strategy} field in \code{\link[=job]{job()}}}
\usage{
strategy(matrix = NULL, `fail-fast` = NULL, `max-parallel` = NULL)
}
\arguments{
\item{matrix}{\code{[list(list(c()))]}
giving the values for each variable for the matrix build.
See \code{\link[=gh_matrix]{gh_matrix()}} for additional options.
Defaults to \code{NULL}.}

\item{`fail-fast`}{\code{[logical()]}
giving whether GitHub should cancel all in-progress jobs if any matrix job fails.
Defaults to \code{NULL}.}

\item{`max-parallel`}{\code{[integer(1)]}
giving the maximum number of jobs to run simultaneously when using a matrix job strategy.}
}
\description{
Create nested list for the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idstrategy}{strategy} field in \code{\link[=job]{job()}}
}
\seealso{
Other syntax: \code{\link{container}},
  \code{\link{gh_matrix}}, \code{\link{ghactions_events}},
  \code{\link{ghactions_vms}}, \code{\link{io}},
  \code{\link{job}}, \code{\link{on}},
  \code{\link{workflow}}
}
\concept{syntax}
