% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_dockerfile}
\alias{use_dockerfile}
\title{Set up a simple \code{Dockerfile}}
\usage{
use_dockerfile(FROM = "rocker/verse:3.5.2")
}
\arguments{
\item{FROM}{\code{[character(1)]}
giving the base docker image.
See details.}
}
\description{
Set up a simple \code{Dockerfile}
}
\details{
Every project to be run on GitHub actions needs a \code{Dockerfile} at the root of the repository.
For many projects, the popular \href{https://hub.docker.com/r/rocker/verse}{verse} image, maintained by the \href{http://rocker-project.org/}{Rocker Project} will suffice; it includes RStudio, the tidyverse, many frequently used packages and system dependencies.
If you need more (or want less), you can always edit your \code{Dockerfile} by hand.
Learn more about \href{https://www.rocker-project.org/use/extending/}{extending images in the context of R} at the Rocker Project.
}
\seealso{
Other setup: \code{\link{edit_workflow}},
  \code{\link{use_ghactions_badge}},
  \code{\link{use_ghactions}}
}
\concept{setup}
