% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{website}
\alias{website}
\title{Render and deploy a website}
\usage{
website(IDENTIFIER = "Render and Deploy RMarkdown Website",
  deploy = list(master = ghpages()))
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action or workflow block.

Used:
\itemize{
\item as an informative label on GitHub.com,
\item in the \code{needs} fields of other \emph{action blocks} to model the workflow graph,
\item in the \code{resolves} fields of other \emph{workflow blocks} to model the workflow graph.
}}

\item{deploy}{\code{[list(1)]}
giving the name of the branch to deploy \emph{from}, and the function to deploy \emph{with}.}
}
\value{
A list as specified in the \code{workflow} argument to \code{\link[=use_ghactions]{use_ghactions()}}.
}
\description{
This workflow renders some Rmarkdown via its (custom) site generator and deploys the result.
Suitable for:
\itemize{
\item \href{https://rmarkdown.rstudio.com/lesson-13.html}{RMarkdown websites}
\item \href{https://bookdown.org}{Bookdown websites}
\item \href{https://bookdown.org/yihui/blogdown/}{Blogdown websites}
\item any other custom site generators that can be called via \code{rmarkdown::render_site()}
}
}
\concept{workflows}
