% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{workflow}
\alias{workflow}
\alias{workflow2hcl}
\title{Create GitHub Actions syntax for \emph{one workflow}.}
\usage{
workflow(IDENTIFIER, on = "push", resolves)

workflow2hcl(l)
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the workflow block.}

\item{on}{\code{[character(1)]}
giving the \href{https://developer.github.com/webhooks/#events}{GitHub Event} on which to trigger the workflow.
Must be one of \link{ghactions_events}.
Defaults to \code{"push"}, in which case the workflow is triggered on every push event.}

\item{resolves}{\code{[character()]}
giving the action(s) to resolve}

\item{l}{\code{[list()]}
giving the named list of HCL fields as returned by \code{\link[=action]{action()}} and \code{\link[=workflow]{workflow()}}.}
}
\value{
\code{[list()]}
A list as specified in the \code{workflow} argument to \code{\link[=use_ghactions]{use_ghactions()}}.
}
\description{
Create GitHub Actions syntax for \emph{one workflow}.
}
\section{Functions}{
\itemize{
\item \code{workflow2hcl}: Convert workflow block to HCL
}}

\examples{
workflow(
  IDENTIFIER = "Run calculation",
  on = "push",
  resolves = "Simple Addition"
)

}
\concept{syntax workflows}
