% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_com.R
\name{check_clean_tree}
\alias{check_clean_tree}
\title{Check \code{git status} for a clean working tree}
\usage{
check_clean_tree(code, dir = getwd())
}
\arguments{
\item{code}{The code to execute.}

\item{dir}{The directory \emph{in} which to execute the code.
Defaults to \code{\link[=getwd]{getwd()}}.}
}
\value{
\code{[character(1)]} The \code{git status} results or \code{TRUE} if no diffs.
}
\description{
Check whether some code will cause changes to \code{git status} in the working directory.
}
\details{
The contents of \code{dir} will be copied to a temporary directory, where a git repository will be initiated and the \code{code} will be executed.
There will never be any changes to \code{dir}.
If \code{dir} or its subdirectories contain a \code{.gitignore}, it will be respected.

This function is modelled \href{https://mllg.github.io/checkmate/articles/checkmate.html}{checkmate}.
}
\concept{prog_com}
\keyword{internal}
