% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_com.R
\name{check_clean_tree}
\alias{check_clean_tree}
\alias{assert_clean_tree}
\title{Check \code{git status} for a clean working tree}
\usage{
check_clean_tree(code = NULL, path = getwd(), before_code = NULL)

assert_clean_tree(code = NULL, path = getwd(), before_code = NULL,
  .var.name = checkmate::vname(code), add = NULL)
}
\arguments{
\item{code}{The code to execute.
Defaults to \code{NULL}.}

\item{path}{The directory in which to execute the code.
Defaults to \code{\link[=getwd]{getwd()}}.}

\item{before_code}{\code{[character(1)]} Giving what happens when the working tree is \emph{already unclean} before \code{code} is evaluated:
\itemize{
\item \code{"stop"} to throw an error,
\item \code{"stash"} to \code{git stash push} all changes before, and \code{git stash pop} them after \code{code} is run (\emph{not recommended}).
Might fail in unexpected ways, including merge conflicts.
\item \code{"commit"} to \code{git add .; git commit -m "commit to cleanup"} all changes before \code{code}is run (\emph{not recommended}).
Might fail in unexpected ways and alter the git history if not run in an isolated container or environment.
\item \code{NULL} (\emph{recommended} default), in which case if \code{is.act()}, then \code{"commit"}, otherwise \code{"stop"}.
}}
}
\value{
\code{[character(1)]} The \code{git status} results or \code{TRUE} if no diffs.
}
\description{
Check whether some code will cause changes to \code{git status} in the working directory.
}
\details{
The contents of \code{path} will be copied to a temporary directory, where a git repository will be initiated and the \code{code} will be executed.
There will never be any changes to \code{path}.
If \code{path} or its subdirectories contain a \code{.gitignore}, it will be respected.

This function is modelled \href{https://mllg.github.io/checkmate/articles/checkmate.html}{checkmate}.
}
\seealso{
Other prog_com: \code{\link{document}}
}
\concept{prog_com}
\keyword{internal}
