% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{checkout}
\alias{checkout}
\title{Create a step to checkout a repository}
\usage{
checkout()
}
\description{
Create a step to checkout a repository
}
\seealso{
Other steps: 
\code{\link{deploy_static}},
\code{\link{install_deps}()},
\code{\link{pkg_dev}},
\code{\link{rscript}()}

Other installation: 
\code{\link{install_deps}()}
}
\concept{installation}
\concept{steps}
