% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{docker_cli}
\alias{docker_cli}
\alias{build_image}
\title{Create \href{https://github.com/actions/docker/tree/aea64bb1b97c42fa69b90523667fef56b90d7cff}{Docker CLI action} to run \href{http://docker.com}{Docker}}
\usage{
docker_cli(IDENTIFIER, needs, args)

build_image(...)
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action or workflow block.

Used:
\itemize{
\item as an informative label on GitHub.com,
\item in the \code{needs} fields of other \emph{action blocks} to model the workflow graph,
\item in the \code{resolves} fields of other \emph{workflow blocks} to model the workflow graph.
}}

\item{needs}{\code{[character()]}
giving the actions (by their \code{IDENTIFIER}s) that must complete successfully before this action will be invoked.
Defaults to \code{NULL} for no upstream dependencies.}

\item{args}{\code{[character()]}
giving the arguments to pass to the action.
Arguments get appended to the last command in \code{ENTRYPOINT}.
Defaults to \code{NULL} for no arguments.}

\item{...}{arguments passed on to other methods, not currently used.}
}
\description{
Thin wrapper around GitHub actions.
}
\details{
These functions are for \strong{advanced users} knowledgeable about GitHub actions.
Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories.
Essentially, they just create lists ready to be ingested by \code{\link[=action2hcl]{action2hcl()}}, which then turns these R lists into valid GitHub actions syntax blocks.

For documentation on these actions, consult their respective \code{README.md}s linked in the below.
Some variants of these action wrappers include sensible defaults for frequent uses in R.

The \code{uses} field is \emph{always} hardcoded to a particular commit or tag of the underlying github action to ensure compatibility.

To render an action block completely from scratch, you can always use the templating function \code{\link[=action]{action()}}.
}
\section{Functions}{
\itemize{
\item \code{build_image}: Build image called \code{repo:latest} from \code{Dockerfile} at repository root
}}

\seealso{
Other actions: \code{\link{document}},
  \code{\link{filter}}, \code{\link{firebase}},
  \code{\link{ghpages}}, \code{\link{install_deps}},
  \code{\link{netlify}}, \code{\link{rscript_byod}},
  \code{\link{rsync}}
}
\concept{actions}
