% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog_com.R
\name{document}
\alias{document}
\title{Roxygenize package and check for inconsistencies}
\usage{
document(dir = getwd())
}
\arguments{
\item{dir}{The directory \emph{in} which to execute the code.
Defaults to \code{\link[=getwd]{getwd()}}.}
}
\description{
Runs \code{\link[devtools:document]{devtools::document()}} and checks whether there are any differences to the working tree.
}
\seealso{
Other prog_com: \code{\link{check_clean_tree}}
}
\concept{prog_com}
\keyword{internal}
