% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{edit_workflow}
\alias{edit_workflow}
\title{Open workflow file for editing}
\usage{
edit_workflow()
}
\value{
Target path, invisibly.
}
\description{
Opens a file for editing in RStudio, if that is the active environment, or
via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise. If the file does not exist, it is
created. If the parent directory does not exist, it is also created.
}
\examples{
\dontrun{
edit_file("DESCRIPTION")
edit_file("~/.gitconfig")
}
}
\seealso{
Other setup: \code{\link{use_dockerfile}},
  \code{\link{use_ghactions_badge}},
  \code{\link{use_ghactions}}
}
\concept{setup}
