% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{filter}
\alias{filter}
\alias{filter_branch}
\title{Create \href{https://github.com/actions/bin/tree/a9036ccda9df39c6ca7e1057bc4ef93709adca5f/filter}{filter action}}
\usage{
filter(IDENTIFIER, needs, args)

filter_branch(needs, branch = "master")
}
\arguments{
\item{branch}{\code{[character(1)]}
giving the name of the branch to filter on.}

\item{...}{arguments passed on to other methods, not currently used.}
}
\description{
Thin wrapper around GitHub actions.
}
\details{
These functions are for \strong{advanced users} knowledgeable about GitHub actions.
Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories.
Essentially, they just create lists ready to be ingested by \code{\link[=make_action_block]{make_action_block()}}, which then turns these R lists into valid GitHub actions syntax blocks.

For documentation on these actions, consult their respective \code{README.md}s linked in the below.
Some variants of these action wrappers include sensible defaults for frequent uses in R.

The \code{uses} field is \emph{always} hardcoded to a particular commit or tag of the underlying github action to ensure compatibility.

To render an action block completely from scratch, you can always use the templating function \code{\link[=make_action_block]{make_action_block()}}.
}
\section{Functions}{
\itemize{
\item \code{filter_branch}: Filter on branch
}}

\seealso{
Other actions: \code{\link{docker_cli}},
  \code{\link{document}}, \code{\link{firebase}},
  \code{\link{ghpages}}, \code{\link{install_deps}},
  \code{\link{netlify}}, \code{\link{rscript_byod}},
  \code{\link{rsync}}
}
\concept{actions}
