% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghactions-package.R
\docType{package}
\name{ghactions}
\alias{ghactions}
\alias{ghactions-package}
\title{ghactions: GitHub actions for R}
\description{
Provides GitHub actions and some R functions to create workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxheld83/ghactions}
  \item Report bugs at \url{https://github.com/maxheld83/ghactions/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (0000-0002-4703-5388)

Other contributors:
\itemize{
  \item Verena Held \email{info@verenaheld.in} [contributor]
}

}
\keyword{internal}
