% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{ghpages}
\alias{ghpages}
\title{Create an action step to deploy via \href{https://pages.github.com}{GitHub Pages}}
\usage{
ghpages(`if` = "github.ref == 'refs/heads/master'",
  name = "Deploy to GitHub Pages", BUILD_DIR = "_site", ...)
}
\arguments{
\item{BUILD_DIR}{\code{[character(1)]}
giving the path relative from your \code{/github/workspace} to the directory to be published.}

\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
}}
}
\description{
\strong{Remember to provide a GitHub personal access token secret named \code{GH_PAT} to the GitHub UI.}
\enumerate{
\item Set up a new PAT.
You can use \code{\link[usethis:browse_github_pat]{usethis::browse_github_pat()}} to get to the right page.
Remember that this PAT is \emph{not} for your local machine, but for GitHub actions.
\item Copy the PAT to your clipboard.
\item Go to the settings of your repository, and paste the PAT as a secret.
The secret must be called \code{GH_PAT}.
}
}
\details{
Wraps the external \href{https://github.com/maxheld83/ghpages/}{ghpages action}.
}
\seealso{
Other steps actions: \code{\link{netlify}},
  \code{\link{rsync}}
}
\concept{steps actions}
