% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{ghpages}
\alias{ghpages}
\title{Create \href{https://github.com/maxheld83/ghpages/tree/v0.1.2}{ghpages action} to deploy to \href{https://pages.github.com}{GitHub Pages}.}
\usage{
ghpages(IDENTIFIER = "Deploy", needs = "Filter on master",
  BUILD_DIR = "_site", env = NULL)
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action or workflow block.

Used:
\itemize{
\item as an informative label on GitHub.com,
\item in the \code{needs} fields of other \emph{action blocks} to model the workflow graph,
\item in the \code{resolves} fields of other \emph{workflow blocks} to model the workflow graph.
}}

\item{needs}{\code{[character()]}
giving the actions (by their \code{IDENTIFIER}s) that must complete successfully before this action will be invoked.
Defaults to \code{NULL} for no upstream dependencies.}

\item{BUILD_DIR}{\code{[character(1)]}
giving the path relative from your \code{/github/workspace} to the directory to be published.}

\item{env}{\code{[list(character(1)]}
giving the environment variables to set in the action's runtime environment.
Defaults to \code{NULL} for no environment variables.}

\item{...}{arguments passed on to other methods, not currently used.}
}
\description{
\strong{Remember to provide \code{GH_PAT} as a secret to the GitHub UI.}
See above docs.

Thin wrapper around GitHub actions.
}
\details{
These functions are for \strong{advanced users} knowledgeable about GitHub actions.
Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories.
Essentially, they just create lists ready to be ingested by \code{\link[=make_action_block]{make_action_block()}}, which then turns these R lists into valid GitHub actions syntax blocks.

For documentation on these actions, consult their respective \code{README.md}s linked in the below.
Some variants of these action wrappers include sensible defaults for frequent uses in R.

The \code{uses} field is \emph{always} hardcoded to a particular commit or tag of the underlying github action to ensure compatibility.

To render an action block completely from scratch, you can always use the templating function \code{\link[=make_action_block]{make_action_block()}}.
}
\seealso{
Other actions: \code{\link{docker_cli}},
  \code{\link{filter}}, \code{\link{netlify}},
  \code{\link{rscript_byod}}, \code{\link{rsync}}
}
\concept{actions}
