% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{ghpages}
\alias{ghpages}
\title{Create \href{https://github.com/maxheld83/ghpages/tree/v0.2.0}{ghpages action} to deploy to \href{https://pages.github.com}{GitHub Pages}}
\usage{
ghpages(IDENTIFIER = "Deploy", needs = "Filter master",
  BUILD_DIR = "_site", env = NULL)
}
\arguments{
\item{BUILD_DIR}{\code{[character(1)]}
giving the path relative from your \code{/github/workspace} to the directory to be published.}

\item{...}{arguments passed on to other methods, not currently used.}
}
\description{
\strong{Remember to provide a GitHub personal access token secret named \code{GH_PAT} to the GitHub UI.}
\enumerate{
\item Set up a new PAT.
You can use \code{\link[usethis:browse_github_pat]{usethis::browse_github_pat()}} to get to the right page.
Remember that this PAT is \emph{not} for your local machine, but for GitHub actions.
\item Copy the PAT to your clipboard.
\item Go to the settings of your repository, and paste the PAT as a secret.
The secret must be called \code{GH_PAT}.
}

For details, see docs of the \href{https://github.com/maxheld83/ghpages/tree/v0.2.0}{ghpages action}.

Thin wrapper around GitHub actions.
}
\details{
These functions are for \strong{advanced users} knowledgeable about GitHub actions.
Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories.
Essentially, they just create lists ready to be ingested by \code{\link[=make_action_block]{make_action_block()}}, which then turns these R lists into valid GitHub actions syntax blocks.

For documentation on these actions, consult their respective \code{README.md}s linked in the below.
Some variants of these action wrappers include sensible defaults for frequent uses in R.

The \code{uses} field is \emph{always} hardcoded to a particular commit or tag of the underlying github action to ensure compatibility.

To render an action block completely from scratch, you can always use the templating function \code{\link[=make_action_block]{make_action_block()}}.
}
\seealso{
Other actions: \code{\link{docker_cli}},
  \code{\link{document}}, \code{\link{filter}},
  \code{\link{firebase}}, \code{\link{install_deps}},
  \code{\link{netlify}}, \code{\link{rscript_byod}},
  \code{\link{rsync}}
}
\concept{actions}
