% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{install_deps}
\alias{install_deps}
\title{Create a step to install R package dependencies}
\usage{
install_deps(name = "Install Package Dependencies", ...)
}
\arguments{
\item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}

\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{working-directory}{\code{[character(1)]}
giving the default working directory.
Defaults to \code{NULL}.}
  \item{continue-on-error}{\code{[logical(1)]}
giving whether to allow a job to pass when this step fails.
Defaults to \code{NULL}.}
  \item{timeout-minutes}{\code{[integer(1)]}
giving the maximum number of minutes to run the step before killing the process.
Defaults to \code{NULL}.}
}}
}
\description{
Installs R package dependencies from a \code{DESCRIPTION} at the repository root.
}
\seealso{
Other steps: \code{\link{checkout}}, \code{\link{ghpages}},
  \code{\link{pkg_dev}}, \code{\link{rscript}}

Other installation: \code{\link{checkout}}
}
\concept{installation}
\concept{steps}
