% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{io}
\alias{io}
\alias{write_workflow}
\alias{read_workflows}
\alias{read_workflow}
\title{Reading and writing GitHub Actions workflow files}
\usage{
write_workflow(x, file = stdout(), ...)

read_workflows(path = ".github/workflows")

read_workflow(file, ...)
}
\arguments{
\item{x}{\code{[list()]}
as created by the workflow functions.}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing}

\item{...}{arguments to \code{\link{as.yaml}}}

\item{path}{\code{[character()]} giving the directory from the repository root where to find GitHub Actions workflows.
Defaults to \code{".github/workflows"}.}
}
\value{
\code{[list()]} of lists from yaml.
}
\description{
Reading and writing GitHub Actions workflow files
}
\details{
It is not necessary to escape characters with special meaning in yaml; the underlying \code{\link[yaml:write_yaml]{yaml::write_yaml()}} does this automatically.

If a workflow is \emph{not} \code{name:}d, the file name will be used as a \code{name: }, as per the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#name}{GitHub Actions documentation}.
}
\section{Functions}{
\itemize{
\item \code{write_workflow}: Write \emph{one} GitHub Actions workflow to file.

\item \code{read_workflows}: Read in \emph{one or more} GitHub Actions workflows from file(s)

\item \code{read_workflow}: Read in \emph{one} GitHub Actions workflow from a file.
}}

\seealso{
Other syntax: \code{\link{ghactions_events}},
  \code{\link{ghactions_vms}}, \code{\link{job}},
  \code{\link{on}}, \code{\link{workflow}}
}
\concept{syntax}
