% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{job}
\alias{job}
\title{Create nested list for \emph{one} \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobs}{job}}
\usage{
job(id, name = NULL, needs = NULL, runs_on = "ubuntu-18.04",
  steps = NULL, timeout_minutes = NULL, strategy = NULL,
  container = NULL)
}
\arguments{
\item{id, name}{\code{[character(1)]}
giving additional options for the job.
Defaults to \code{NULL}.}

\item{needs}{\code{[character()]}
giving the jobs that must complete successfully before this job is run.
Defaults to \code{NULL} for no dependencies.}

\item{runs_on}{\code{[character(1)]}
giving the type of virtual host machone to run the job on.
Defaults to \code{"ubuntu-18.04"}.
Must be one of \link{ghactions_events}.}

\item{steps}{\code{[list()]}
giving an \emph{unnamed} list of steps, with each element as returned by \code{\link[=step]{step()}}.
Defaults to \code{NULL}.}

\item{timeout_minutes}{\code{[integer(1)]}
giving the maximum number of minutes to let a workflow run before GitHub automatically cancels it.
Defaults to \code{NULL}.}

\item{strategy}{\code{[list()]}
giving a named list as returned by \code{\link[=strategy]{strategy()}}.
Defaults to \code{NULL}.}

\item{container}{\code{[character(1)]}/\code{[list()]}
giving a published container image.
For advanced options, use \code{\link[=container]{container()}}.
Defaults to \code{NULL}.}
}
\description{
Create nested list for \emph{one} \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobs}{job}
}
\seealso{
Other syntax: \code{\link{gh_matrix}},
  \code{\link{ghactions_events}},
  \code{\link{ghactions_vms}}, \code{\link{io}},
  \code{\link{on}}, \code{\link{strategy}},
  \code{\link{workflow}}
}
\concept{syntax}
