% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{make_blocks}
\alias{make_blocks}
\alias{make_workflow_block}
\title{Create GitHub Actions syntax blocks}
\usage{
make_workflow_block(IDENTIFIER, on = "push", resolves)
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action or workflow block.

Used:
\itemize{
\item as an informative label on GitHub.com,
\item in the \code{needs} fields of other \emph{action blocks} to model the workflow graph,
\item in the \code{resolves} fields of other \emph{workflow blocks} to model the workflow graph.
}}

\item{on}{\code{[character(1)]}
giving the \href{https://developer.github.com/webhooks/#events}{GitHub Event} on which to trigger the workflow.
Must be one of \link{ghactions_events}.
Defaults to \code{"push"}, in which case the workflow is triggered on every push event.}

\item{resolves}{\code{[character()]}
giving the action(s) to invoke.}
}
\value{
\code{[character(1)]}
}
\description{
Creates the syntax building blocks for GitHub actions: workflows and actions.
For details on the syntax and arguments, see \href{https://developer.github.com/actions/creating-workflows/workflow-configuration-options/}{here}.
}
\section{Functions}{
\itemize{
\item \code{make_workflow_block}: Write GitHub Actions syntax for \emph{one} workflow block.
}}

\examples{
make_workflow_block(
  IDENTIFIER = "Run calculation",
  on = "push",
  resolves = "Simple Addition"
)

}
\seealso{
Other syntax: \code{\link{action}},
  \code{\link{ghactions_events}}

Other syntax: \code{\link{action}},
  \code{\link{ghactions_events}}
}
\concept{syntax}
