% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{netlify}
\alias{netlify}
\title{Create an action step to deploy to \href{https://www.netlify.com}{Netlify}}
\usage{
netlify(name = "Deploy to Netlify",
  `if` = "github.ref == 'refs/heads/master'", dir = "$DEPLOY_PATH",
  prod = TRUE, with = NULL, env = NULL, site, ...)
}
\arguments{
\item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}

\item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}

\item{dir}{\code{[character(1)]}
giving the path relative from your \code{/github/workspace} to the directory to be published.}

\item{prod}{\code{[logical(1)]}
giving whether the deploy should be to production.}

\item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}

\item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}

\item{site}{\code{[character(1)]}
giving a site ID to deploy to.}

\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{working-directory}{\code{[character(1)]}
giving the default working directory.
Defaults to \code{NULL}.}
  \item{continue-on-error}{\code{[logical(1)]}
giving whether to allow a job to pass when this step fails.
Defaults to \code{NULL}.}
  \item{timeout-minutes}{\code{[integer(1)]}
giving the maximum number of minutes to run the step before killing the process.
Defaults to \code{NULL}.}
}}
}
\description{
\strong{Remember to provide \code{NETLIFY_AUTH_TOKEN} and \code{NETLIFY_SITE_ID} (optional) as secrets to the GitHub UI.}
}
\details{
Wraps the external \href{https://github.com/netlify/actions}{netlify cli action}.
}
\seealso{
Other steps: \code{\link{checkout}},
  \code{\link{firebase}}, \code{\link{ghpages}},
  \code{\link{install_deps}}, \code{\link{pkg_dev}},
  \code{\link{rscript}}, \code{\link{rsync}}

Other actions: \code{\link{firebase}},
  \code{\link{ghpages}}, \code{\link{rsync}}

Other deployment: \code{\link{firebase}},
  \code{\link{ghpages}}, \code{\link{rsync}}
}
\concept{actions}
\concept{deployment}
\concept{steps}
