% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{pkg_dev}
\alias{pkg_dev}
\alias{rcmd_check}
\alias{covr}
\title{CI/CD steps for a package at the repository root}
\usage{
rcmd_check(name = "Check Package")

covr(name = "Run Code Coverage")
}
\arguments{
\item{name}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
}
\description{
CI/CD steps for a package at the repository root
}
\section{Functions}{
\itemize{
\item \code{rcmd_check}: \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}

\item \code{covr}: \code{\link[covr:codecov]{covr::codecov()}}
}}

\seealso{
Other steps: 
\code{\link{checkout}()},
\code{\link{deploy_static}},
\code{\link{install_deps}()},
\code{\link{rscript}()}
}
\concept{pkg_development}
\concept{steps}
