% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{pkg_dev}
\alias{pkg_dev}
\alias{rcmd_check}
\alias{covr}
\title{CI/CD steps for a package at the repository root}
\usage{
rcmd_check(name = "Check Package")

covr(name = "Run Code Coverage")
}
\arguments{
\item{name}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}

\item{...}{
  Arguments passed on to \code{\link[=step]{step}}
  \describe{
    \item{\code{id}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{if}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{shell}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{with}}{\verb{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
    \item{\code{env}}{\verb{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
    \item{\code{working-directory}}{\verb{[character(1)]}
giving the default working directory.
Defaults to \code{NULL}.}
    \item{\code{continue-on-error}}{\verb{[logical(1)]}
giving whether to allow a job to pass when this step fails.
Defaults to \code{NULL}.}
    \item{\code{timeout-minutes}}{\verb{[integer(1)]}
giving the maximum number of minutes to run the step before killing the process.
Defaults to \code{NULL}.}
  }}
}
\description{
CI/CD steps for a package at the repository root
}
\section{Functions}{
\itemize{
\item \code{rcmd_check}: \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}

\item \code{covr}: \code{\link[covr:codecov]{covr::codecov()}}
}}

\seealso{
Other steps: 
\code{\link{checkout}()},
\code{\link{deploy_static}},
\code{\link{install_deps}()},
\code{\link{rscript}()}
}
\concept{pkg_development}
\concept{steps}
