% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{rscript}
\alias{rscript}
\title{Create a step to run \link[utils:Rscript]{utils::Rscript}}
\usage{
rscript(options = NULL, expr = NULL, file = NULL, args = NULL, ...)
}
\arguments{
\item{options}{a list of options, all beginning with \samp{--}.  These
    can be any of the options of the standard \R front-end, and also those
    described in the details.}

\item{expr}{\R expression(s), properly quoted.}

\item{file}{the name of a file containing \R commands.  \samp{-}
    indicates \file{stdin}.}

\item{args}{arguments to be passed to the script in \code{file}.}

\item{...}{
  Arguments passed on to \code{\link[=step]{step}}
  \describe{
    \item{\code{id}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{if}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{name}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{uses}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{shell}}{\verb{[character(1)]}
giving additional options for the step.
Multiline strings are not supported.
Defaults to \code{NULL}.}
    \item{\code{with}}{\verb{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
    \item{\code{env}}{\verb{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
    \item{\code{working-directory}}{\verb{[character(1)]}
giving the default working directory.
Defaults to \code{NULL}.}
    \item{\code{continue-on-error}}{\verb{[logical(1)]}
giving whether to allow a job to pass when this step fails.
Defaults to \code{NULL}.}
    \item{\code{timeout-minutes}}{\verb{[integer(1)]}
giving the maximum number of minutes to run the step before killing the process.
Defaults to \code{NULL}.}
  }}
}
\description{
Create a step to run \link[utils:Rscript]{utils::Rscript}
}
\seealso{
Other steps: 
\code{\link{checkout}()},
\code{\link{deploy_static}},
\code{\link{install_deps}()},
\code{\link{pkg_dev}}
}
\concept{script}
\concept{steps}
