% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{rsync}
\alias{rsync}
\title{Create an action step to deploy via \href{https://rsync.samba.org}{Rsync} over SSH}
\usage{
rsync(HOST_NAME, HOST_IP, HOST_FINGERPRINT, src, user, dest,
  `if` = "github.ref == 'refs/heads/master'", env = NULL,
  with = NULL, ...)
}
\arguments{
\item{HOST_NAME}{\code{[character(1)]}
giving the name of the server you wish to deploy to, such as \code{foo.example.com}.}

\item{HOST_IP}{\code{[character(1)]}
giving the IP of the server you wish to deploy to, such as \code{111.111.11.111}.}

\item{HOST_FINGERPRINT}{\code{[character(1)]}
giving the fingerprint of the server you wish to deploy to, can have different formats.}

\item{src}{\code{[character(1)]}
giving the source directory, relative path \emph{from} \code{/github/workspace} \strong{without trailing slash}.}

\item{user}{\code{[character(1)]}
giving the user at the target \code{HOST_NAME}.}

\item{dest}{\code{[character(1)]}
giving the directory from the root of the \code{HOST_NAME} target to write to.}

\item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}

\item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}

\item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}

\item{...}{Arguments passed on to \code{step}
\describe{
  \item{id}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{if}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{name}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}
  \item{with}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}
  \item{working-directory}{\code{[character(1)]}
giving the default working directory.
Defaults to \code{NULL}.}
  \item{continue-on-error}{\code{[logical(1)]}
giving whether to allow a job to pass when this step fails.
Defaults to \code{NULL}.}
  \item{timeout-minutes}{\code{[integer(1)]}
giving the maximum number of minutes to run the step before killing the process.
Defaults to \code{NULL}.}
}}
}
\description{
\strong{Remember to provide \code{SSH_PRIVATE_KEY} and \code{SSH_PUBLIC_KEY} as secrets to the GitHub UI.}.
}
\details{
Wraps the external \href{https://github.com/maxheld83/rsync/}{rsync action}.
}
\seealso{
Other steps: \code{\link{checkout}},
  \code{\link{firebase}}, \code{\link{ghpages}},
  \code{\link{install_deps}}, \code{\link{netlify}},
  \code{\link{pkg_dev}}, \code{\link{rscript}}

Other actions: \code{\link{firebase}},
  \code{\link{ghpages}}, \code{\link{netlify}}

Other deployment: \code{\link{firebase}},
  \code{\link{ghpages}}, \code{\link{netlify}}
}
\concept{actions}
\concept{deployment}
\concept{steps}
