% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{step}
\alias{step}
\title{Create nested list for \emph{one} \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobs}{job}}
\usage{
step(name = NULL, id = NULL, `if` = NULL, uses = NULL,
  run = NULL, shell = NULL, with = NULL, env = NULL,
  `working-directory` = NULL, `continue-on-error` = NULL,
  `timeout-minutes` = NULL)
}
\arguments{
\item{id, if, name, uses, shell}{\code{[character(1)]}
giving additional options for the stop.
Multiline strings are not supported.
Defaults to \code{NULL}.}

\item{run}{\code{[character()]}
giving commands to run.
Will be turned into a multiline string.
Defaults to \code{NULL}.}

\item{with, env}{\code{[list()]}
giving a named list of additional parameters.
Defaults to \code{NULL}.}

\item{working-directory}{\code{[character(1)]}
giving the default working directory.
Defaults to \code{NULL}.}

\item{continue-on-error}{\code{[logical(1)]}
giving whether to allow a job to pass when this step fails.
Defaults to \code{NULL}.}

\item{timeout-minutes}{\code{[integer(1)]}
giving the maximum number of minutes to run the step before killing the process.
Defaults to \code{NULL}.}
}
\description{
Create nested list for \emph{one} \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobs}{job}
}
