% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_ghactions}
\alias{use_ghactions}
\title{Workflow automation with GitHub Actions}
\usage{
use_ghactions(workflow = website())
}
\arguments{
\item{workflow}{\code{[list(list())]}
A named list of as created by one of the \code{\link[=workflow]{workflow()}} functions.}
}
\value{
A logical vector indicating if file was modified.
}
\description{
Sets up workflow automation, including continuous integration and deployment (CI/CD) for different kinds of R projects on GitHub actions.
This function
\itemize{
\item Picks a set of sensible defaults for your project,
\item transforms a list into the GitHub actions syntax.
\item writes it out to \code{.github/workflows/} in your repository.
}
}
\examples{
\dontrun{
use_ghactions(workflow = website())
}
}
\seealso{
Other setup: \code{\link{edit_workflow}},
  \code{\link{is_github_actions}},
  \code{\link{use_ghactions_badge}}
}
\concept{setup}
