% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{use_ghactions}
\alias{use_ghactions}
\title{Workflow automation with GitHub Actions}
\usage{
use_ghactions(workflow = workflows$website$rmarkdown$fau())
}
\arguments{
\item{workflow}{\code{[list()]}
A named list of blocks nested as:
\itemize{
\item arguments to \code{make_workflow_block} as a named list \emph{and}
\item \code{actions}, which in turn comprises the
\itemize{
\item arguments to \code{make_action_block} as a named list.
Defaults ti \code{NULL}, in which case one of the \code{workflows} functions is chosen based on the configuration files in your repository.
}
}}
}
\description{
Sets up workflow automation, including continuous integration and deployment (CI/CD) for a different kinds of R projects on GitHub actions.
This function
\itemize{
\item Picks a set of sensible defaults for your project.
\item Transforms a list of workflow and action blocks into the GitHub actions syntax.
\item Adds a \code{.github/main.workflow} file to your repository.
}
}
\examples{
\dontrun{
use_ghactions(workflow = workflows$website$rmarkdown)
}
}
