% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_ghactions_badge}
\alias{use_ghactions_badge}
\title{README badges}
\usage{
use_ghactions_badge()
}
\description{
Add markdown syntax for README badge, see \code{\link[usethis:use_badge]{usethis::use_badge()}}.
}
\seealso{
Other setup: \code{\link{use_dockerfile}},
  \code{\link{use_ghactions}}
}
\concept{setup}
