% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{website}
\alias{website}
\title{Render and deploy an \href{https://rmarkdown.rstudio.com}{rmarkdown} project}
\usage{
website(name = "Render and Deploy RMarkdown Website",
  deploy = list(ghpages()), on = c("push", "pull_request"))
}
\arguments{
\item{name}{\code{[character(1)]}
giving the \href{https://help.github.com/en/articles/workflow-syntax-for-github-actions#name}{name} of the workflow.
Defaults to \code{NULL}, for no name, in which case GitHub will use the file name.}

\item{deploy}{\code{[list(1)]}
giving a list of deploy functions.}

\item{on}{\code{[character()]}
giving the \href{https://help.github.com/en/articles/events-that-trigger-workflows}{GitHub Event} on which to trigger the workflow.
Must be a subset of \link{ghactions_events}.
Defaults to \code{"push"}, in which case the workflow is triggered on every push event.
Can also be a named list as returned by \code{\link[=on]{on()}} for additional filters.}
}
\description{
This workflow renders some Rmarkdown via its (custom) site generator and deploys the result.
Suitable for:
\itemize{
\item \href{https://rmarkdown.rstudio.com/lesson-13.html}{RMarkdown websites}
\item \href{https://bookdown.org}{Bookdown websites}
\item \href{https://bookdown.org/yihui/blogdown/}{Blogdown websites} (\strong{experimental})
}
}
\details{
\code{\link[rmarkdown:render_site]{rmarkdown::render_site()}} returns the directory to which outputs have been rendered.
This workflow saves that directory in the environment variable \code{DEPLOY_PATH}, which is the default expected by \code{\link[=ghpages]{ghpages()}} and other deployment methods.
}
