% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{website}
\alias{website}
\alias{website_rmarkdown}
\title{Render and deploy R projects to static hosting}
\usage{
website(IDENTIFIER, fun, deploy = list(master = ghpages()))

website_rmarkdown(...)
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action or workflow block.

Used:
\itemize{
\item as an informative label on GitHub.com,
\item in the \code{needs} fields of other \emph{action blocks} to model the workflow graph.
\item in the \code{resolves} fields of other \emph{workflow blocks} to model the workflow graph.
}}

\item{fun}{\code{[character(1)]}
giving the function call.}

\item{deploy}{\code{[list(1)]}
giving the name of the branch to deploy \emph{from}, and the function to deploy \emph{with}.}

\item{...}{arguments passed on to other methods, not currently used.}
}
\value{
A list as specified in the \code{workflow} argument to \code{\link[=use_ghactions]{use_ghactions()}}.
}
\description{
These functions generate workflows suitable for R \code{website} projects.
You can use some of the below defaults, or call an arbitrary R function that generates static assets.
}
\section{Functions}{
\itemize{
\item \code{website_rmarkdown}: \href{https://rmarkdown.rstudio.com/lesson-13.html}{RMarkdown websites}
}}

\concept{workflows}
