% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{website}
\alias{website}
\title{Render and deploy R projects to static hosting}
\usage{
website(IDENTIFIER = "Render and Deploy",
  .f = "rmarkdown::render_site()", static_dir = "_site",
  deploy = NULL)
}
\arguments{
\item{IDENTIFIER}{\code{[character(1)]}
giving the name of the action or workflow block.
Used:
\itemize{
\item as informative label on GitHub.com,
\item in the \code{needs} fields of other \emph{action blocks} to express dependencies.
\item in the \code{resolves} fields of other \emph{workflow blocks} to express dependencies.
}}

\item{.f}{\code{[character(1)]}
giving the render function, typically something like \code{rmarkdown::render_site()}.}

\item{static_dir}{\code{[character(1)]}
giving the path to the generated static assets from the repository root.
Typically something like \code{_site/} for \code{rmarkdown::render_site()}.}

\item{deploy}{\code{[list()]}
giving the name of the branch to deploy \emph{from}, and the function to deploy \emph{with}.}
}
\value{
A list as specified in the \code{workflow} argument to \code{\link[=use_ghactions]{use_ghactions()}}.
}
\description{
These functions generate workflows suitable for R \code{website} projects, such as:
\itemize{
\item \href{https://rmarkdown.rstudio.com/lesson-13.html}{RMarkdown websites}
\item \href{http://bookdown.org}{Bookdown websites}
\item \href{https://bookdown.org/yihui/blogdown/}{Blogdown websites}
\item and other similar projects with an R static site generator.
}
}
\concept{workflows}
